+++ lib/src/t_notification_overlay.dart
@@ -13,8 +13,10 @@ class TNotificationOverlay {
   /// Displays a success notification.
   static void success({
     required BuildContext context,
-    required Widget title,
-    Widget? subTitle,
+    required String title,
+    String? subTitle,
+    Widget? action,
+    bool? sticky,
     Duration duration = const Duration(seconds: 3),
     double spacing = 10,
     double height = 100,
@@ -25,6 +27,8 @@ class TNotificationOverlay {
       context: context,
       title: title,
       subTitle: subTitle,
+      action: action,
+      sticky: sticky,
       type: NotificationType.success,
       duration: duration,
       spacing: spacing,
@@ -37,8 +41,10 @@ class TNotificationOverlay {
   /// Displays an error notification.
   static void error({
     required BuildContext context,
-    required Widget title,
-    Widget? subTitle,
+    required String title,
+    String? subTitle,
+    Widget? action,
+    bool? sticky,
     Duration duration = const Duration(seconds: 3),
     double spacing = 10,
     double height = 100,
@@ -49,6 +55,8 @@ class TNotificationOverlay {
       context: context,
       title: title,
       subTitle: subTitle,
+      action: action,
+      sticky: sticky,
       type: NotificationType.error,
       duration: duration,
       spacing: spacing,
@@ -61,8 +69,10 @@ class TNotificationOverlay {
   /// Displays a warning notification.
   static void warning({
     required BuildContext context,
-    required Widget title,
-    Widget? subTitle,
+    required String title,
+    String? subTitle,
+    Widget? action,
+    bool? sticky,
     Duration duration = const Duration(seconds: 3),
     double spacing = 10,
     double height = 100,
@@ -73,6 +83,8 @@ class TNotificationOverlay {
       context: context,
       title: title,
       subTitle: subTitle,
+      action: action,
+      sticky: sticky,
       type: NotificationType.warning,
       duration: duration,
       spacing: spacing,
@@ -85,8 +97,10 @@ class TNotificationOverlay {
   /// Displays an info notification.
   static void info({
     required BuildContext context,
-    required Widget title,
-    Widget? subTitle,
+    required String title,
+    String? subTitle,
+    Widget? action,
+    bool? sticky,
     Duration duration = const Duration(seconds: 3),
     double spacing = 10,
     double height = 100,
@@ -97,6 +111,8 @@ class TNotificationOverlay {
       context: context,
       title: title,
       subTitle: subTitle,
+      action: action,
+      sticky: sticky,
       type: NotificationType.info,
       duration: duration,
       spacing: spacing,
@@ -108,18 +124,29 @@ class TNotificationOverlay {
 
   /// Displays a notification with a specific type, title, and message.
   ///
-  /// [title] is the title of the notification.
-  /// [subTitle] is the detailed message of the notification.
-  /// [type] determines the style (success, error, warning, info).
-  /// [duration] defines how long the notification will remain visible (default is 3 seconds).
-  /// [spacing] allows customization of the space between notifications (default is 10.0).
-  /// [position] defines the corner where notifications will appear (default is topRight).
+  /// [title] is the main title of the notification (required).
+  /// [subTitle] is the optional detailed message of the notification.
+  /// [action] is a customizable widget for user actions (e.g., buttons).
+  /// [type] determines the style of the notification (success, error, warning, info).
+  /// [duration] defines how long the notification remains visible. Ignored if [sticky] is true.
+  /// [sticky] makes the notification stay on screen indefinitely (default is false).
+  /// [spacing] specifies the space between stacked notifications (default is 10.0).
+  /// [height] sets the height of the notification widget.
+  /// [width] specifies the width of the notification widget.
+  /// [titleColor], [messageColor], [backgroundColor], and [borderColor] customize colors.
+  /// [borderRadius] sets the corner radius of the notification box.
+  /// [iconColor] customizes the notification icon color.
+  /// [paddingVertical] and [paddingHorizontal] adjust internal padding.
+  /// [slideInDirection] and [slideOutDirection] control the slide animations.
+  /// [position] defines the notification's position on the screen (e.g., topRight, bottomLeft).
   static void show({
     required BuildContext context,
-    required Widget title,
-    Widget? subTitle,
+    required String title,
+    String? subTitle,
+    Widget? action,
     NotificationType type = NotificationType.info,
     Duration duration = const Duration(seconds: 3),
+    bool? sticky,
     double spacing = 10,
     double height = 100,
     double? width,
@@ -142,27 +169,20 @@ class TNotificationOverlay {
     late final OverlayEntry overlayEntry;
 
     // Add a global key to control the position animation.
-    final GlobalKey<NotificationAnimationPositionState> key =
-        GlobalKey<NotificationAnimationPositionState>();
+    final GlobalKey<NotificationAnimationPositionState> key = GlobalKey<NotificationAnimationPositionState>();
 
     overlayEntry = OverlayEntry(
       builder: (context) {
         // Determine the alignment based on user-specified position.
-        final bool isTop = position == NotificationPosition.topLeft ||
-            position == NotificationPosition.topRight;
-        final bool isLeft = position == NotificationPosition.topLeft ||
-            position == NotificationPosition.bottomLeft;
+        final bool isTop = position == NotificationPosition.topLeft || position == NotificationPosition.topRight;
+        final bool isLeft = position == NotificationPosition.topLeft || position == NotificationPosition.bottomLeft;
 
         return AnimatedPositioned(
           duration: Duration(milliseconds: 300),
           left: isLeft ? 16.0 : null,
           right: isLeft ? null : 16.0,
-          top: isTop
-              ? _calculateOffset(overlayEntry, height, spacing, isTop: true)
-              : null,
-          bottom: isTop
-              ? null
-              : _calculateOffset(overlayEntry, height, spacing, isTop: false),
+          top: isTop ? _calculateOffset(overlayEntry, height, spacing, isTop: true) : null,
+          bottom: isTop ? null : _calculateOffset(overlayEntry, height, spacing, isTop: false),
           child: NotificationAnimationPosition(
             key: key,
             slideInDirection: slideInDirection,
@@ -170,11 +190,8 @@ class TNotificationOverlay {
             child: TNotificationWidget(
               title: title,
               subTitle: subTitle,
-              onClose: () {
-                key.currentState?.animateOut(() {
-                  _removeNotification(overlayEntry);
-                });
-              },
+              action: action,
+              sticky: sticky,
               type: type,
               width: width,
               paddingHorizontal: paddingHorizontal,
@@ -186,6 +203,11 @@ class TNotificationOverlay {
               borderRadius: borderRadius,
               iconColor: iconColor,
               duration: duration,
+              onClose: () {
+                key.currentState?.animateOut(() {
+                  _removeNotification(overlayEntry);
+                });
+              },
             ),
           ),
         );
@@ -197,15 +219,15 @@ class TNotificationOverlay {
     overlay?.insert(overlayEntry);
 
     // Automatically remove the notification after the specified duration.
-    Future.delayed(
+    if(sticky == null || !sticky) {
+      Future.delayed(
         duration,
-        () => key.currentState
-            ?.animateOut(() => _removeNotification(overlayEntry)));
+        () => key.currentState?.animateOut(() => _removeNotification(overlayEntry)),
+      );
+    }
   }
 
-  static double _calculateOffset(
-      OverlayEntry entry, double height, double spacing,
-      {required bool isTop}) {
+  static double _calculateOffset(OverlayEntry entry, double height, double spacing, {required bool isTop}) {
     final index = _notifications.indexOf(entry);
     final totalOffset = index * (height + spacing);
     return totalOffset + (isTop ? height : 0.0); // Add margin for padding.