+++ lib/src/t_notification_widget.dart
@@ -1,14 +1,17 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:iconsax/iconsax.dart';
 
 import '../t_overlay_notification.dart';
 
 class TNotificationWidget extends StatefulWidget {
-  final Widget title;
-  final Widget? subTitle;
+  final String title;
+  final String? subTitle;
+  final Widget? action;
   final NotificationType type;
   final VoidCallback onClose;
+  final bool? sticky;
 
   final Color? backgroundColor;
   final Color? borderColor;
@@ -27,6 +30,7 @@ class TNotificationWidget extends StatefulWidget {
     required this.onClose,
     required this.duration,
     this.subTitle,
+    this.action,
     this.backgroundColor,
     this.borderColor,
     this.titleColor,
@@ -36,6 +40,7 @@ class TNotificationWidget extends StatefulWidget {
     this.borderRadius,
     this.paddingVertical,
     this.paddingHorizontal,
+    this.sticky,
     super.key,
   });
 
@@ -51,7 +56,7 @@ class _TNotificationWidgetState extends State<TNotificationWidget> {
   void initState() {
     super.initState();
     _progressValue = 1.0;
-    _startSliderCountdown();
+    if (widget.sticky == null || !widget.sticky!) _startSliderCountdown();
   }
 
   void _startSliderCountdown() {
@@ -82,25 +87,22 @@ class _TNotificationWidgetState extends State<TNotificationWidget> {
   @override
   Widget build(BuildContext context) {
     final Map<NotificationType, Color> defaultTypeColors = {
-      NotificationType.success: const Color(0xFF4CAF50),
-      NotificationType.error: const Color(0xFFF44336),
-      NotificationType.warning: const Color(0xFFFFC107),
-      NotificationType.info: const Color(0xFF2196F3),
+      NotificationType.success: const Color(0xFF8FB9A8),
+      NotificationType.error: const Color(0xFFD8737F),
+      NotificationType.warning: const Color(0xFFFCBB6D),
+      NotificationType.info: const Color(0xFF475C7A),
     };
 
     final Map<NotificationType, IconData> typeIcons = {
-      NotificationType.success: Icons.check,
-      NotificationType.error: Icons.error,
-      NotificationType.warning: Icons.warning,
-      NotificationType.info: Icons.info,
+      NotificationType.success: Iconsax.tick_circle,
+      NotificationType.error: Iconsax.close_circle,
+      NotificationType.warning: Iconsax.warning_2,
+      NotificationType.info: Iconsax.info_circle,
     };
 
-    final Color effectiveBackgroundColor = widget.backgroundColor ??
-        defaultTypeColors[widget.type]!.withValues(alpha: 0.2);
-    final Color effectiveBorderColor =
-        widget.borderColor ?? defaultTypeColors[widget.type]!;
-    final Color effectiveIconColor =
-        widget.iconColor ?? defaultTypeColors[widget.type]!;
+    final Color effectiveBackgroundColor = widget.backgroundColor ?? defaultTypeColors[widget.type]!;
+    final Color effectiveBorderColor = widget.borderColor ?? defaultTypeColors[widget.type]!;
+    final Color effectiveIconColor = widget.iconColor ?? Colors.white;
     final double effectiveWidth = widget.width ?? 350.0;
     final double effectiveBorderRadius = widget.borderRadius ?? 16.0;
     final double effectivePaddingVertical = widget.paddingVertical ?? 8.0;
@@ -113,12 +115,8 @@ class _TNotificationWidgetState extends State<TNotificationWidget> {
         // padding: EdgeInsets.only(top: effectivePaddingVertical),
         decoration: BoxDecoration(
           color: effectiveBackgroundColor,
-          border:
-              Border(left: BorderSide(color: effectiveBorderColor, width: 1)),
-          borderRadius: BorderRadius.only(
-            topRight: Radius.circular(effectiveBorderRadius),
-            bottomRight: Radius.circular(effectiveBorderRadius),
-          ),
+          border: Border(left: BorderSide(color: effectiveBorderColor, width: 1)),
+          borderRadius: BorderRadius.circular(effectiveBorderRadius),
         ),
         width: effectiveWidth,
         child: Column(
@@ -127,23 +125,29 @@ class _TNotificationWidgetState extends State<TNotificationWidget> {
           children: [
             /// Value
             ListTile(
-              title: widget.title,
-              subtitle: widget.subTitle,
-              contentPadding:
-                  EdgeInsets.symmetric(horizontal: effectivePaddingHorizontal),
-              leading: Icon(typeIcons[widget.type],
-                  size: 24.0, color: effectiveIconColor),
+              title: Text(widget.title, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16, color: Colors.white)),
+              subtitle: widget.action != null
+                  ? Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        if (widget.subTitle != null)
+                          Text(widget.subTitle!, style: TextStyle(fontWeight: FontWeight.normal, fontSize: 14, color: Colors.white)),
+                        widget.action!,
+                      ],
+                    )
+                  : widget.subTitle != null
+                      ? Text(widget.subTitle!, style: TextStyle(fontWeight: FontWeight.normal, fontSize: 14, color: Colors.white))
+                      : null,
+              contentPadding: EdgeInsets.symmetric(horizontal: effectivePaddingHorizontal),
+              leading: Icon(typeIcons[widget.type], size: 24.0, color: effectiveIconColor),
               trailing: IconButton(
-                icon: const Icon(Icons.close, size: 18.0),
+                icon: const Icon(Icons.close, size: 18.0, color: Colors.white),
                 onPressed: widget.onClose,
               ),
             ),
 
             /// Progress Indicator
-            LinearProgressIndicator(
-                minHeight: 1,
-                value: _progressValue,
-                color: effectiveBorderColor),
+            LinearProgressIndicator(minHeight: 1, value: _progressValue, color: effectiveBorderColor),
           ],
         ),
       ),